Unison/requirements.txt
fastapi==0.114.2
motor==3.5.1
passlib==1.7.4
pydantic[email]==2.9.1
pydantic_core==2.23.3
pydantic_settings==2.5.2
pymongo==4.8.0
pyotp==2.9.0
python_jose==3.3.0
pytz==2024.2
uvicorn==0.30.6
python-multipart==0.0.9
